requirements.txt
fastapi==0.85.1
python-multipart==0.0.5
tensorflow-cpu==2.10.0
uvicorn==0.18.3
gunicorn==20.1.0